+++ Hazel/src/Platform/System/FileTracker.h
@@ -13,12 +13,12 @@ namespace Hazel {
 	class FileTracker
 	{
 	public:
-		static unsigned long long GetCurrentlyOpenFilesCount();//The number of active file handles
-		static unsigned long long GetOpenedFilesCount();//Total ever opened / ever closed
-		static unsigned long long GetClosedFilesCount();
+		static uint64_t GetCurrentlyOpenFilesCount();//The number of active file handles
+		static uint64_t GetOpenedFilesCount();//Total ever opened / ever closed
+		static uint64_t GetClosedFilesCount();
 
-		static unsigned long long GetOpenedFilesSec();//The number of files opened/closed in the last second
-		static unsigned long long GetClosedFilesSec();
+		static uint64_t GetOpenedFilesSec();//The number of files opened/closed in the last second
+		static uint64_t GetClosedFilesSec();
 		static inline CountTracker& GetTracker() { return m_FileTracker; }
 		static inline std::map<FILE*, std::string>& GetOpenFilePaths() { return m_OpenFiles; }
 		static inline std::set<std::string>& GetClosedFilePaths() { return m_ClosedFiles; }
